monitoring_automation_v2/requirements.txt
boto3
elasticsearch
pyyaml
json_logging
gitpython